+++ tests/test_experiment.py
@@ -1,0 +1,2 @@
+def test_experiment(py_or_json):
+    print(py_or_json)
