+++ tests/test_json.py
@@ -40,13 +40,6 @@ def test_model():
     assert v.validate_json(input_str) == ({'field_a': '123', 'field_b': 1}, {'field_b', 'field_a'})
 
 
-def test_list():
-    v = SchemaValidator({'type': 'list', 'items': {'type': 'int'}, 'title': 'TestModel'})
-    # language=json
-    input_str = '[1, 2, "3"]'
-    assert v.validate_json(input_str) == [1, 2, 3]
-
-
 def test_float_no_remainder():
     v = SchemaValidator({'type': 'int'})
     assert v.validate_json('123.0') == 123
