+++ tests/validators/test_bool.py
@@ -1,0 +1,72 @@
+import re
+
+import pytest
+
+from pydantic_core import SchemaValidator, ValidationError
+
+from ..conftest import Err
+
+
+@pytest.mark.parametrize(
+    'input_value,expected',
+    [
+        (False, False),
+        (True, True),
+        (0, False),
+        (1, True),
+        ('yes', True),
+        ('no', False),
+        ('true', True),
+        ('false', False),
+        (
+            'cheese',
+            Err(
+                'Value must be a valid boolean, '
+                "unable to interpret input [kind=bool_parsing, input_value='cheese', input_type=str]"
+            ),
+        ),
+    ],
+)
+def test_bool(py_or_json, input_value, expected):
+    v = py_or_json({'type': 'bool'})
+    if isinstance(expected, Err):
+        with pytest.raises(ValidationError, match=re.escape(expected.message)):
+            v.validate_test(input_value)
+    else:
+        assert v.validate_test(input_value) == expected
+
+
+def test_bool_strict():
+    v = SchemaValidator({'type': 'bool', 'strict': True})
+    assert v.validate_python(True) is True
+    error_message = "Value must be a valid boolean [kind=bool_type, input_value='true', input_type=str]"
+    with pytest.raises(ValidationError, match=re.escape(error_message)):
+        v.validate_python('true')
+
+
+def test_bool_error():
+    v = SchemaValidator({'type': 'bool'})
+
+    with pytest.raises(ValidationError) as exc_info:
+        v.validate_python('wrong')
+
+    assert str(exc_info.value) == (
+        '1 validation error for bool\n'
+        '  Value must be a valid boolean, '
+        "unable to interpret input [kind=bool_parsing, input_value='wrong', input_type=str]"
+    )
+    assert exc_info.value.errors() == [
+        {
+            'kind': 'bool_parsing',
+            'loc': [],
+            'message': 'Value must be a valid boolean, unable to interpret input',
+            'input_value': 'wrong',
+        }
+    ]
+
+
+def test_bool_repr():
+    v = SchemaValidator({'type': 'bool'})
+    assert repr(v) == 'SchemaValidator(name="bool", validator=BoolValidator)'
+    v = SchemaValidator({'type': 'bool', 'strict': True})
+    assert repr(v) == 'SchemaValidator(name="strict-bool", validator=StrictBoolValidator)'
