+++ tests/validators/test_dict.py
@@ -6,11 +6,11 @@ import pytest
 from pydantic_core import SchemaValidator, ValidationError
 
 
-def test_dict():
-    v = SchemaValidator({'type': 'dict', 'keys': {'type': 'int'}, 'values': {'type': 'int'}})
-    assert v.validate_python({'1': 2, '3': 4}) == {1: 2, 3: 4}
-    v = SchemaValidator({'type': 'dict', 'strict': True, 'keys': {'type': 'int'}, 'values': {'type': 'int'}})
-    assert v.validate_python({'1': 2, '3': 4}) == {1: 2, 3: 4}
+def test_dict(py_or_json):
+    v = py_or_json({'type': 'dict', 'keys': {'type': 'int'}, 'values': {'type': 'int'}})
+    assert v.validate_test({'1': 2, '3': 4}) == {1: 2, 3: 4}
+    v = py_or_json({'type': 'dict', 'strict': True, 'keys': {'type': 'int'}, 'values': {'type': 'int'}})
+    assert v.validate_test({'1': 2, '3': 4}) == {1: 2, 3: 4}
 
 
 def test_dict_any_value():
