+++ tests/validators/test_model.py
@@ -1,8 +1,6 @@
-import re
-
 import pytest
 
-from pydantic_core import SchemaError, SchemaValidator, ValidationError
+from pydantic_core import SchemaValidator, ValidationError
 
 
 def test_simple():
@@ -215,176 +213,5 @@ def test_validate_assignment_allow_extra_validate():
             'loc': ['other_field'],
             'message': 'Value must be a valid integer, unable to parse string as an integer',
             'input_value': 'xyz',
-        }
-    ]
-
-
-def test_model_class():
-    class MyModel:
-        # this is not required, but it avoids `__fields_set__` being included in `__dict__`
-        __slots__ = '__dict__', '__fields_set__'
-        field_a: str
-        field_b: int
-
-    v = SchemaValidator(
-        {
-            'type': 'model-class',
-            'class': MyModel,
-            'model': {'type': 'model', 'fields': {'field_a': {'type': 'str'}, 'field_b': {'type': 'int'}}},
-        }
-    )
-    assert repr(v).startswith('SchemaValidator(name="MyModel", validator=ModelClassValidator {\n')
-    m = v.validate_python({'field_a': 'test', 'field_b': 12})
-    assert isinstance(m, MyModel)
-    assert m.field_a == 'test'
-    assert m.field_b == 12
-    assert m.__fields_set__ == {'field_a', 'field_b'}
-    assert m.__dict__ == {'field_a': 'test', 'field_b': 12}
-
-
-def test_model_class_setattr():
-    setattr_calls = []
-
-    class MyModel:
-        field_a: str
-
-        def __setattr__(self, key, value):
-            setattr_calls.append((key, value))
-            # don't do anything
-
-    m1 = MyModel()
-    m1.foo = 'bar'
-    assert not hasattr(m1, 'foo')
-    assert setattr_calls == [('foo', 'bar')]
-    setattr_calls.clear()
-
-    v = SchemaValidator(
-        {'type': 'model-class', 'class': MyModel, 'model': {'type': 'model', 'fields': {'field_a': {'type': 'str'}}}}
-    )
-    m = v.validate_python({'field_a': 'test'})
-    assert isinstance(m, MyModel)
-    # assert m.field_a == 'test'
-    # assert m.__fields_set__ == {'field_a'}
-    assert setattr_calls == []
-
-
-def test_model_class_root_validator():
-    class MyModel:
-        pass
-
-    def f(input_value, *, validator, **kwargs):
-        output = validator(input_value)
-        return str(output)
-
-    v = SchemaValidator(
-        {
-            'title': 'Test',
-            'type': 'function',
-            'mode': 'wrap',
-            'function': f,
-            'field': {
-                'type': 'model-class',
-                'class': MyModel,
-                'model': {'type': 'model', 'fields': {'field_a': {'type': 'str'}}},
-            },
-        }
-    )
-    m = v.validate_python({'field_a': 'test'})
-    assert isinstance(m, str)
-    assert 'test_model_class_root_validator.<locals>.MyModel' in m
-
-
-def test_model_class_bad_model():
-    class MyModel:
-        pass
-
-    with pytest.raises(SchemaError, match=re.escape("model-class expected a 'model' schema, got 'str'")):
-        SchemaValidator({'type': 'model-class', 'class': MyModel, 'model': {'type': 'str'}})
-
-
-def test_model_class_not_type():
-    with pytest.raises(SchemaError, match=re.escape("TypeError: 'int' object cannot be converted to 'PyType'")):
-        SchemaValidator({'type': 'model-class', 'class': 123})
-
-
-def test_model_class_instance_direct():
-    class MyModel:
-        __slots__ = '__dict__', '__fields_set__'
-        field_a: str
-
-        def __init__(self):
-            self.field_a = 'init'
-
-    v = SchemaValidator(
-        {'type': 'model-class', 'class': MyModel, 'model': {'type': 'model', 'fields': {'field_a': {'type': 'str'}}}}
-    )
-    m1 = v.validate_python({'field_a': 'test'})
-    assert isinstance(m1, MyModel)
-    assert m1.field_a == 'test'
-    assert m1.__fields_set__ == {'field_a'}
-
-    m2 = MyModel()
-    m3 = v.validate_python(m2)
-    assert m2 == m3
-    assert m3.field_a == 'init'
-
-
-def test_model_class_instance_subclass():
-    class MyModel:
-        __slots__ = '__dict__', '__fields_set__'
-        field_a: str
-
-        def __init__(self):
-            self.field_a = 'init_a'
-
-    class MySubModel(MyModel):
-        field_b: str
-
-        def __init__(self):
-            super().__init__()
-            self.field_b = 'init_b'
-
-    v = SchemaValidator(
-        {'type': 'model-class', 'class': MyModel, 'model': {'type': 'model', 'fields': {'field_a': {'type': 'str'}}}}
-    )
-
-    m2 = MySubModel()
-    assert m2.field_a
-    m3 = v.validate_python(m2)
-    assert m2 != m3
-    assert m3.field_a == 'init_a'
-    assert not hasattr(m3, 'field_b')
-
-
-def test_model_class_strict():
-    class MyModel:
-        def __init__(self):
-            self.field_a = 'init_a'
-            self.field_b = 'init_b'
-
-    v = SchemaValidator(
-        {
-            'type': 'model-class',
-            'strict': True,
-            'class': MyModel,
-            'model': {'type': 'model', 'fields': {'field_a': {'type': 'str'}, 'field_b': {'type': 'int'}}},
-        }
-    )
-    m = MyModel()
-    m2 = v.validate_python(m)
-    assert isinstance(m, MyModel)
-    assert m is m2
-    assert m.field_a == 'init_a'
-    # note that since dict validation was not run here, there has been no check this is an int
-    assert m.field_b == 'init_b'
-    with pytest.raises(ValidationError) as exc_info:
-        v.validate_python({'field_a': 'test', 'field_b': 12})
-    assert exc_info.value.errors() == [
-        {
-            'kind': 'model_type',
-            'loc': [],
-            'message': 'Value must be an instance of MyModel',
-            'input_value': {'field_a': 'test', 'field_b': 12},
-            'context': {'class_name': 'MyModel'},
         }
     ]
